Move selectors to store
