+++ client/src/schema/SchemaSidebar.tsx
@@ -12,12 +12,8 @@ import Sidebar from '../common/Sidebar';
 import SpinKitCube from '../common/SpinKitCube';
 import Text from '../common/Text';
 import Tooltip from '../common/Tooltip';
-import { useSelectedConnectionId } from '../stores/editor-actions';
-import {
-  loadSchemaInfo,
-  toggleSchemaItem,
-  useSchema,
-} from '../stores/editor-actions';
+import { loadSchemaInfo, toggleSchemaItem } from '../stores/editor-actions';
+import { useSchema, useSelectedConnectionId } from '../stores/editor-store';
 import getSchemaList from './getSchemaList';
 import styles from './SchemaSidebar.module.css';
 import searchSchemaInfo from './searchSchemaInfo';
